requirements.txt
ansi2html==1.8.0
anyio==4.0.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astor==0.8.1
asttokens==2.4.0
astunparse==1.6.3
async-lru==2.0.4
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.9.0
bleach==6.0.0
blinker==1.4
brotlipy==0.7.0
catkin-pkg==0.5.2
chardet==5.2.0
comm==0.1.4
ConfigArgParse==1.7
contourpy==1.1.0
cycler==0.11.0
dash==2.13.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.7.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.8.0
docutils==0.20.1
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.18.0
filelock==3.12.3
Flask==2.2.5
flatbuffers==23.5.26
fonttools==4.42.1
fqdn==1.5.1
google-auth-oauthlib==1.0.0
h5py==2.10.0
importlib-resources==6.0.1
ipykernel==6.25.2
ipython==8.12.2
ipywidgets==8.1.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
keras==2.13.1
kiwisolver==1.4.5
libclang==16.0.6
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==3.0.1
mkl-fft==1.3.6
mkl-service==2.4.0
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.7.0
nest-asyncio==1.5.7
networkx @ file:///C:/b/abs_e6gi1go5op/croot/networkx_1690562046966/work
notebook_shim==0.2.3
numpy==1.23.5
oauthlib @ file:///C:/b/abs_44riqpe5gg/croot/oauthlib_1679489644705/work
open3d==0.17.0
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
overrides==7.4.0
packaging @ file:///C:/b/abs_28t5mcoltc/croot/packaging_1693575224052/work
pandas==0.25.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
pickleshare==0.7.5
Pillow==10.0.0
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
plotly==5.16.1
pooch @ file:///tmp/build/80754af9/pooch_1623324770023/work
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pybullet==3.2.5
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.16.1
PyJWT @ file:///C:/ci/pyjwt_1657529430378/work
pyOpenSSL @ file:///C:/b/abs_08f38zyck4/croot/pyopenssl_1690225407403/work
pyparsing==3.0.9
pyreadline==2.1
PySocks @ file:///C:/ci/pysocks_1605287845585/work
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-ignite==0.4.12
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
requests @ file:///C:/b/abs_316c2inijk/croot/requests_1690400295842/work
requests-oauthlib==1.3.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rospkg==1.5.0
rpds-py==0.10.2
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
scipy==1.4.0
Send2Trash==1.8.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.2
sympy @ file:///C:/b/abs_95fbf1z7n6/croot/sympy_1668202411612/work
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboard-plugin-wit @ file:///C:/tf/b/tensorboard-plugin-wit_1660162132996/work/tensorboard_plugin_wit-1.8.1-py3-none-any.whl
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor @ file:///C:/b/abs_16qe7jhw7n/croot/termcolor_1668084642458/work
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.3.0
urllib3 @ file:///C:/b/abs_889_loyqv4/croot/urllib3_1686163174463/work
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.2
Werkzeug @ file:///C:/b/abs_8578rs2ra_/croot/werkzeug_1679489759009/work
widgetsnbextension==4.0.8
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306167264/work
wrapt @ file:///C:/Windows/Temp/abs_7c3dd407-1390-477a-b542-fd15df6a24085_diwiza/croots/recipe/wrapt_1657814452175/work
yarl @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_e5nlunspj6/croots/recipe/yarl_1661437086516/work
zipp @ file:///C:/b/abs_b9jfdr908q/croot/zipp_1672387552360/work
